+++ .vs/VSWorkspaceState.json
@@ -1,11 +1,7 @@
 {
   "ExpandedNodes": [
-    "",
-    "\\Jongeun",
-    "\\Jongeun\\BlackJack",
-    "\\Jongeun\\BlackJack\\BlackJack",
-    "\\Jongeun\\BlackJack\\Debug"
+    ""
   ],
-  "SelectedNode": "\\Jongeun\\BlackJack\\BlackJack\\BlackJack.cpp",
+  "SelectedNode": "\\BlackJack.sln",
   "PreviewInSolutionExplorer": false
 }
