+++ Jongeun/BlackJack/BlackJack.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.9.34723.18
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BlackJack", "BlackJack\BlackJack.vcxproj", "{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Test", "Test\Test.vcxproj", "{D80627CC-DE1A-459A-8113-2B4478BE2295}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -21,6 +23,14 @@ Global
 		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Release|x64.Build.0 = Release|x64
 		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Release|x86.ActiveCfg = Release|Win32
 		{4EAF5DB2-67AB-49D2-8D0A-2B83E769276C}.Release|x86.Build.0 = Release|Win32
+		{D80627CC-DE1A-459A-8113-2B4478BE2295}.Debug|x64.ActiveCfg = Debug|x64
+		{D80627CC-DE1A-459A-8113-2B4478BE2295}.Debug|x64.Build.0 = Debug|x64
+		{D80627CC-DE1A-459A-8113-2B4478BE2295}.Debug|x86.ActiveCfg = Debug|Win32
+		{D80627CC-DE1A-459A-8113-2B4478BE2295}.Debug|x86.Build.0 = Debug|Win32
+		{D80627CC-DE1A-459A-8113-2B4478BE2295}.Release|x64.ActiveCfg = Release|x64
+		{D80627CC-DE1A-459A-8113-2B4478BE2295}.Release|x64.Build.0 = Release|x64
+		{D80627CC-DE1A-459A-8113-2B4478BE2295}.Release|x86.ActiveCfg = Release|Win32
+		{D80627CC-DE1A-459A-8113-2B4478BE2295}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
