+++ Jongeun/BlackJack/BlackJack/Random.h
@@ -1,6 +1,9 @@
 #ifndef RANDOM_MT_H
 #define RANDOM_MT_H
 
+/**
+header only library
+**/
 #include <chrono>
 #include <random>
