+++ Jongeun/BlackJack/Test/x64/Debug/Test.log
@@ -1,0 +1,2 @@
+﻿  test.cpp
+  Test.vcxproj -> C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\x64\Debug\Test.exe
