+++ Jongeun/BlackJack/Test/x64/Debug/Test.tlog/Cl.items.tlog
@@ -1,0 +1,1 @@
+C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\Test\test.cpp;C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\Test\x64\Debug\test.obj
