+++ Jongeun/BlackJack/Test/x64/Debug/Test.tlog/Test.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.38.33130:TargetPlatformVersion=10.0.22621.0:
+Debug|x64|C:\Users\user\source\repos\SKKUREAL\miniGame\Jongeun\BlackJack\|
